08a_autogen_conversational_ai/00_single_agent/chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.4.8.1",
    "autogen-ext[openai]>=0.4.8.1",
    "chainlit>=2.2.1",
    "python-dotenv>=1.0.1",
]
